Move tests for device IO to separate file
